backend/requirements.txt
asgiref==3.8.1
blinker==1.9.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
datetime-truncate==1.1.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.4
django-admin-charts==1.3.1
django-admin-interface==0.26.0
django-bower==5.2.0
django-colorfield==0.11.0
django-cors-headers==4.6.0
django-debug-toolbar==4.3.0
django-import-export==3.2.0
django-logentry-admin==1.1.0
django-memoize==2.3.1
django-multiselectfield==0.1.12
django-nvd3==0.10.1
django-rest-swagger==2.2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
Faker==24.14.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
gunicorn==23.0.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.2.0
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.1.2
packaging==24.0
pika==1.3.2
pillow==10.3.0
pluggy==1.5.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
pyTelegramBotAPI==4.21.0
pytest==8.3.4
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
setuptools==69.5.1
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
stripe==11.2.0
tablib==3.6.1
text-unidecode==1.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.1.3
xlrd==2.0.1
xlwt==1.3.0
